ETL/4_tables_bigquery/requirements_api.txt
functions-framework==3.*
google-cloud-storage
google-cloud-bigquery
pandas
pyarrow
db-dtypes
